Add Spring JPA with entities. Refactoring project.
