+++ .gitignore
@@ -1,3 +1,4 @@
 /.idea/
 *.iml
 /target/
+/src/main/resources/application.properties
