--- src/main/java/IMVC/IController.java
@@ -1,5 +1,0 @@
-package IMVC;
-
-public interface IController {
-    boolean tryLogin(byte[] loginParams);
-}
