--- src/main/java/IMVC/IModel.java
@@ -1,7 +1,0 @@
-package IMVC;
-
-import java.util.Map;
-
-public interface IModel {
-    boolean login(Map<String, String> params);
-}
