--- src/main/java/IMVC/IView.java
@@ -1,5 +1,0 @@
-package IMVC;
-
-public interface IView {
-    void init();
-}
