--- src/main/java/model/Factories/CircleFactory.java
@@ -1,19 +1,0 @@
-package model.Factories;
-
-import model.Shapes.Circle;
-import model.Shapes.Shape;
-
-import java.util.List;
-
-public class CircleFactory implements IShapeFactory {
-
-    @Override
-    public Shape createFigure(List<Double> params) {
-        return new Circle(params);
-    }
-
-    @Override
-    public Shape createFigure(String _id, int id, List<Double> params) {
-        return new Circle(_id, id, params);
-    }
-}
