--- src/main/java/model/IActions/IMovable.java
@@ -1,5 +1,0 @@
-package model.IActions;
-
-public interface IMovable {
-    void move (double x, double y);
-}
