--- src/main/java/model/IActions/IRollable.java
@@ -1,5 +1,0 @@
-package model.IActions;
-
-public interface IRollable {
-    void roll(double angle);
-}
