--- src/main/java/model/IActions/IScalebale.java
@@ -1,7 +1,0 @@
-package model.IActions;
-
-public interface IScalebale {
-    default void changeSize(double scaleFactor) {
-
-    }
-}
