--- src/main/java/model/Shapes/ShapeTypes.java
@@ -1,18 +1,0 @@
-package model.Shapes;
-
-public enum ShapeTypes {
-
-    TRIANGLE ("Треугольник"),
-    RECTANGLE("Прямоугольник"),
-    CIRCLE("Круг");
-
-    private String ruName;
-
-    ShapeTypes(String name) {
-        this.ruName = name;
-    }
-
-    public String getRuName() {
-        return ruName;
-    }
-}
