+++ src/main/java/shapes/old/IActions/IMovable.java
@@ -1,0 +1,5 @@
+package shapes.old.IActions;
+
+public interface IMovable {
+    void move (double x, double y);
+}
