+++ src/main/java/shapes/old/IActions/IRollable.java
@@ -1,0 +1,5 @@
+package shapes.old.IActions;
+
+public interface IRollable {
+    void roll(double angle);
+}
