+++ src/main/java/shapes/old/IActions/IScalebale.java
@@ -1,0 +1,7 @@
+package shapes.old.IActions;
+
+public interface IScalebale {
+    default void changeSize(double scaleFactor) {
+
+    }
+}
